requirements.txt
alabaster==0.7.8
anaconda-client==1.4.0
anaconda-navigator==1.3.1
apiai==1.0.2
appnope==0.1.0
appscript==1.0.1
argcomplete==1.0.0
astroid==1.4.7
astropy==1.2.1
Babel==2.3.3
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.4.1
bitarray==0.8.1
blaze==0.10.1
blinker==1.4
bokeh==0.12.0
boto==2.40.0
Bottleneck==1.1.0
bson==0.4.3
cdecimal==2.3
cffi==1.6.0
chest==0.2.3
click==6.6
cloudpickle==0.2.1
clyent==1.2.2
colorama==0.3.7
conda==4.2.12
conda-build==1.21.3
configobj==5.0.6
configparser==3.5.0b2
contextlib2==0.5.3
cryptography==1.4
cycler==0.10.0
Cython==0.24
cytoolz==0.8.0
dask==0.10.0
datashape==0.5.2
decorator==4.0.10
dill==0.2.5
Django==1.10.5
docutils==0.12
docxtpl==0.3.1
dynd==0.7.3.dev1
enum34==1.1.6
et-xmlfile==1.0.1
Faker==0.7.7
fastcache==1.0.2
feedgenerator==1.8
Flask==0.11.1
Flask-Cors==2.1.2
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.0.5
gevent==1.1.1
greenlet==0.4.10
grin==1.2.1
h5py==2.6.0
HeapDict==1.0.0
hickle==2.1.0
hyperas==0.2
hyperopt==0.0.2
idna==2.1
image==1.5.5
imagesize==0.7.1
ipaddress==1.0.16
ipykernel==4.3.1
ipython==4.2.0
ipython-genutils==0.1.0
ipywidgets==4.1.1
itsdangerous==0.24
jdcal==1.2
jedi==0.9.0
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.3.0
jupyter-console==4.1.1
jupyter-core==4.1.0
Keras==1.1.0
Lasagne==0.2.dev1
lazy-object-proxy==1.2.1
llvmlite==0.11.0
locket==0.2.0
lxml==3.6.0
Markdown==2.6.6
MarkupSafe==0.23
matplotlib==1.5.3
mistune==0.7.2
mock==2.0.0
mpmath==0.19
multipledispatch==0.4.8
nb-anacondacloud==1.1.0
nb-conda==1.1.0
nb-conda-kernels==1.0.3
nbconvert==4.2.0
nbformat==4.0.1
nbpresent==3.0.2
networkx==1.11
nltk==3.2.1
nose==1.3.7
notebook==4.2.1
numba==0.26.0
numexpr==2.6.0
numpy==1.11.2
odo==0.5.0
openpyxl==2.3.2
pandas==0.18.1
partd==0.3.4
path.py==0.0.0
pathlib2==2.1.0
patsy==0.4.1
pbr==1.10.0
pelican==3.6.3
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.7.2
Pillow==3.2.0
ply==3.8
progressbar==2.3
protobuf==3.1.0.post1
psutil==4.3.0
ptyprocess==0.5.1
py==1.4.31
pyasn1==0.1.9
PyAudio==0.2.7
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pycurl==7.43.0
pydot==1.1.0
pyflakes==1.2.3
Pygments==2.1.3
pylint==1.5.4
pymongo==3.3.1
pyOpenSSL==16.0.0
pyparsing==2.1.4
pytest==2.9.2
python-dateutil==2.5.3
python-docx==0.8.6
pytz==2016.4
PyYAML==3.12
pyzmq==15.2.0
QtAwesome==0.3.3
qtconsole==4.2.1
QtPy==1.1.2
redis==2.10.5
requests==2.10.0
requirements==0.1
rope==0.9.4
ruamel-yaml===-VERSION
scikit-image==0.12.3
scikit-learn==0.17.1
scipy==0.18.1
selenium==3.0.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
Sphinx==1.4.1
sphinx-rtd-theme==0.1.9
spyder==3.0.0
SQLAlchemy==1.0.13
statsmodels==0.6.1
sympy==1.0
tables==3.2.2
tensorflow==0.12.1
terminado==0.6
Theano==0.9.0.dev2
toolz==0.8.0
tornado==4.3
tqdm==4.11.2
traitlets==4.2.1
unicodecsv==0.14.1
Unidecode==0.4.19
Wand==0.4.4
Werkzeug==0.11.10
wrapt==1.10.8
xlrd==1.0.0
XlsxWriter==0.9.2
xlwings==0.7.2
xlwt==1.1.2
